@@ -20,26 +20,33 @@
         }
     },
     "require": {
-        "php": "^7.4|^8.0",
-        "react/socket": "^0.8",
-        "react/datagram": "^1.2",
-        "graylog2/gelf-php": "^1.5",
-        "symfony/process": "^4.1|^5.4",
-        "ext-json": "*"
+        "php": "^8.1",
+        "react/socket": "^1.12",
+        "react/datagram": "^1.9",
+        "graylog2/gelf-php": "^2.0",
+        "symfony/process": "^5.4|^6.0",
+        "ext-json": "*",
+        "ext-zlib": "*"
     },
     "require-dev": {
-        "keboola/coding-standard": ">=7.0.2",
-        "phpstan/phpstan": "^1.5",
-        "phpstan/phpstan-phpunit": "^1.0.0",
-        "phpunit/phpunit": "^9.3",
         "ext-sockets": "*",
-        "ext-zlib": "*"
+        "keboola/coding-standard": ">=14.0",
+        "phpstan/phpstan": "^1.9",
+        "phpstan/phpstan-phpunit": "^1.3",
+        "phpstan/phpstan-symfony": "^1.2",
+        "phpunit/phpunit": "^9.5",
+        "sempro/phpunit-pretty-print": "^1.4",
+        "symfony/dotenv": "^5.4|^6.0"
     },
     "scripts": {
         "tests": "phpunit --coverage-clover build/logs/clover.xml --coverage-xml=build/logs/coverage-xml --log-junit=build/logs/phpunit.junit.xml",
         "phpstan": "phpstan analyse --no-progress -c phpstan.neon",
         "phpcs": "phpcs --extensions=php src tests",
+        "phpcbf": "phpcbf --extensions=php src tests",
         "build": [
+            "@phpcs",
+            "@phpstan",
+            "@tests",
             "@tests"
         ],
         "ci": [
@@ -50,6 +57,7 @@
     "config": {
         "sort-packages": true,
         "process-timeout": 3600,
+        "lock": false,
         "allow-plugins": {
             "dealerdirect/phpcodesniffer-composer-installer": true
         }
